kuring-71 [수정] 여러 화면에서 공통으로 쓰이는 Callback을 :common:ui_util로 옮김
